Refactor Font BaseSize Heuristic and Update Models
- Refactored `calculate_base_size` to accept an `Iterable` and process it iteratively for memory efficiency.
- Implemented rounding of font sizes to 1 decimal place before counting to prevent float precision fragmentation.
- Updated `calculate_base_size` to return a `float` instead of an `int`.
- Updated `CodexMeta` model in `engine/codex_engine/models.py` to change `base_size` type to `float`.
- Updated unit tests in `engine/tests/test_heuristics.py` to cover float precision, return type, and generator support.
- Verified all changes by running the test suite with `poetry run pytest`.

I initially had some trouble locating the specific pull request review tools as they weren't standard bash commands, but I was able to successfully retrieve and address the feedback.

+++ engine/codex_engine/models.py
@@ -57,7 +57,7 @@ class CodexMeta(BaseModel):
     title: str
     author: str
     description: Optional[str] = None
-    base_size: int
+    base_size: float
     seo: CodexSEO = Field(default_factory=CodexSEO)
